[WIP] split out helper functions
